Asset Refactor

+++ Modules/Asset/routes/web.php
@@ -8,34 +8,34 @@ use Modules\Asset\Http\Controllers\RepairController;
 
 Route::middleware(['auth', 'tenant'])->prefix('dashboard/assets')->name('assets.')->group(function () {
 
-    Route::get('/', [AssetController::class, 'assetIndex'])->name('index');
-    Route::get('/{asset}/details', [AssetController::class, 'showAssetDetails'])->name('details');
+    Route::get('/', [AssetController::class, 'assetIndex'])->name('index')->middleware('permission:asset,view');
+    Route::get('/{asset}/details', [AssetController::class, 'showAssetDetails'])->name('details')->middleware('permission:asset,view');
 
-    Route::get('/add', [AssetController::class, 'showAssetAdd'])->name('add');
-    Route::post('/add', [AssetController::class, 'storeAssets'])->name('store');
+    Route::get('/add', [AssetController::class, 'showAssetAdd'])->name('add')->middleware('permission:asset,manage');
+    Route::post('/add', [AssetController::class, 'storeAssets'])->name('store')->middleware('permission:asset,manage');
 
-    Route::get('/{asset}/edit', [AssetController::class, 'showAssetEdit'])->name('edit');
-    Route::put('/{asset}/edit', [AssetController::class, 'AssetUpdate'])->name('update');
+    Route::get('/{asset}/edit', [AssetController::class, 'showAssetEdit'])->name('edit')->middleware('permission:asset,manage');
+    Route::put('/{asset}/edit', [AssetController::class, 'AssetUpdate'])->name('update')->middleware('permission:asset,manage');
 
-    Route::delete('/{asset}/delete', [AssetController::class, 'destroy'])->name('delete');
+    Route::delete('/{asset}/delete', [AssetController::class, 'destroy'])->name('delete')->middleware('permission:asset,manage');
 
-    Route::get('/api/{assetType}/assets', [AssetController::class, 'getAssets'])->name('asset.api');
-    Route::post('/import', [AssetImportController::class, 'import'])->name('import');
-    Route::get('/template', [AssetImportController::class, 'template'])->name('template');
-    Route::get('/export', [AssetImportController::class, 'export'])->name('export');
+    Route::get('/api/{assetType}/assets', [AssetController::class, 'getAssets'])->name('asset.api')->middleware('permission:asset,view');
+    Route::post('/import', [AssetImportController::class, 'import'])->name('import')->middleware('permission:asset,manage');
+    Route::get('/template', [AssetImportController::class, 'template'])->name('template')->middleware('permission:asset,manage');
+    Route::get('/export', [AssetImportController::class, 'export'])->name('export')->middleware('permission:asset,manage');
 });
 
 Route::middleware(['auth', 'tenant'])->prefix('dashboard/assettypes')->name('assettypes.')->group(function () {
-    Route::get('/', [AssetTypeController::class, 'assetTypeIndex'])->name('index');
-    Route::get('/{assetType}/details', [AssetTypeController::class, 'assetTypeDetails'])->name('details');
+    Route::get('/', [AssetTypeController::class, 'assetTypeIndex'])->name('index')->middleware('permission:asset,view');
+    Route::get('/{assetType}/details', [AssetTypeController::class, 'assetTypeDetails'])->name('details')->middleware('permission:asset,view');
 
-    Route::get('/add', [AssetTypeController::class, 'showAssetTypeAddForm'])->name('add');
-    Route::post('/add', [AssetTypeController::class, 'storeAssetType'])->name('store');
+    Route::get('/add', [AssetTypeController::class, 'showAssetTypeAddForm'])->name('add')->middleware('permission:asset,manage');
+    Route::post('/add', [AssetTypeController::class, 'storeAssetType'])->name('store')->middleware('permission:asset,manage');
 
-    Route::get('/{assetType}/edit', [AssetTypeController::class, 'showAssetTypeEditForm'])->name('edit');
-    Route::put('/{assetType}/edit', [AssetTypeController::class, 'updateAssetType'])->name('update');
+    Route::get('/{assetType}/edit', [AssetTypeController::class, 'showAssetTypeEditForm'])->name('edit')->middleware('permission:asset,manage');
+    Route::put('/{assetType}/edit', [AssetTypeController::class, 'updateAssetType'])->name('update')->middleware('permission:asset,manage');
 
-    Route::delete('/{assetType}/delete/', [AssetTypeController::class, 'destroy'])->name('delete');
+    Route::delete('/{assetType}/delete/', [AssetTypeController::class, 'destroy'])->name('delete')->middleware('permission:asset,manage');
 });
 
 Route::middleware(['auth', 'tenant'])->prefix('dashboard/repairs')->name('repairs.')->group(function () {
@@ -51,4 +51,4 @@ Route::middleware(['auth', 'tenant'])->prefix('dashboard/repairs')->name('repair
 
     Route::post('/{repair}/cancel', [RepairController::class, 'cancel'])->name('cancel');
     Route::post('/{repair}/complete', [RepairController::class, 'complete'])->name('complete');
-});
+})->middleware('permission:asset,repair');
